apps/trainer/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# ML and Data
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
joblib==1.3.2

# NLP
nltk==3.8.1

# Environment
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
structlog==24.1.0
